requirements.txt
Django==2.2.10
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
